Moved parse command tests to its own file. Added some checks to check for logged errors.
